+++ C#Web/McBonalds/Models/Produto.cs
@@ -4,6 +4,6 @@ namespace McBonalds.Models
     {
         public ulong Id{get;set;}
         public string Nome{get;set;}
-        public float Preco{get;set;}
+        public double Preco{get;set;}
     }
 }
